working on shop module and product catalog management / products & categories CRUD by admin and listing model for customers / add Product/Variant/Category write entities / Product management admin service / Category management admin service / also a new entity called StockChange added for tracking stock history of all products

+++ src/main/java/dev/hspl/hspl2shop/shop/model/impl/jpa/entity/StockChangeJpaEntity.java
@@ -36,7 +36,7 @@ public class StockChangeJpaEntity {
     private boolean increased; // false=decreased
 
     @Column(name = "stock")
-    private int stockItems; // variant stock items before change
+    private short stockItems; // variant stock items before change
 
     @Column(name = "count")
     private short count;
